+++ com.ca.lsp.cobol/lsp-service-cobol/src/main/java/com/ca/lsp/cobol/model/ConfigurationSettingsStorable.java
@@ -15,10 +15,11 @@
 package com.ca.lsp.cobol.model;
 
 import lombok.Data;
+import lombok.Value;
 
 import java.util.List;
 
-@Data
+@Value
 public class ConfigurationSettingsStorable {
 
   private Object profiles;